+++ flow.h
@@ -3,15 +3,15 @@
 #include <queue>
 using namespace std;
 
-bool bfs(vector<vector<int>> &residualGraph, int source, int sink, vector<int> &parent){
+bool bfs(vector<vector<double>> &residualGraph, double source, double sink, vector<double> &parent){
     vector<bool> visited(residualGraph.size(), false);
-    queue<int> q;
+    queue<double> q;
     q.push(source);
     visited[source] = true;
     parent[source] = -1;
 
     while(!q.empty()){
-        int currentNode = q.front();
+        double currentNode = q.front();
         q.pop();
 
         for(int i = 0; i < residualGraph.size(); i++){
@@ -27,8 +27,8 @@ bool bfs(vector<vector<int>> &residualGraph, int source, int sink, vector<int> &
 }
 
 // In a adjacency matrix, check if there are negative values. If there are turn them to cero and add the value to its mirror as abs(value)
-vector<vector<int>> treatGraph(vector<vector<int>> graph){
-    vector<vector<int>> newGraph = graph;
+vector<vector<double>> treatGraph(vector<vector<double>> graph){
+    vector<vector<double>> newGraph = graph;
     for(int i = 0; i < graph.size(); i++){
         for(int j = 0; j < graph.size(); j++){
             if(graph[i][j] < 0){
@@ -42,30 +42,30 @@ vector<vector<int>> treatGraph(vector<vector<int>> graph){
 }
 
 
-int fordFulkerson(vector<vector<int>> graph, int origen, int destino){
+double fordFulkerson(vector<vector<double>> graph, double origen, double destino){
 
     graph = treatGraph(graph);
 
-    int n = graph.size();
-    vector<vector<int>> residualGraph(n, vector<int>(n, 0));
+    double n = graph.size();
+    vector<vector<double>> residualGraph(n, vector<double>(n, 0));
     for(int i = 0; i < n; i++){
         for(int j = 0; j < n; j++){
             residualGraph[i][j] = graph[i][j];
         }
     }
 
-    vector<int> parent(n, -1);
-    int maxFlow = 0;
+    vector<double> parent(n, -1);
+    double maxFlow = 0;
 
     while(bfs(residualGraph, origen, destino, parent)){
-        int pathFlow = INT_MAX;
+        double pathFlow = INT_MAX;
         for(int v = destino; v != origen; v = parent[v]){
-            int u = parent[v];
+            double u = parent[v];
             pathFlow = min(pathFlow, residualGraph[u][v]);
         }
 
         for(int v = destino; v != origen; v = parent[v]){
-            int u = parent[v];
+            double u = parent[v];
             residualGraph[u][v] -= pathFlow;
             residualGraph[v][u] += pathFlow;
         }
